PNO code in ccenergy working again, this time as a separate class
