+++ psi4/src/psi4/cc/cclambda/L1.cc
@@ -38,7 +38,6 @@
 #include "psi4/libqt/qt.h"
 #include "MOInfo.h"
 #include "Params.h"
-#include "cclambda.h"
 #include "Local.h"
 #define EXTERN
 #include "globals.h"
@@ -46,7 +45,6 @@
 namespace psi {
 namespace cclambda {
 
-void pno_filter_T1(dpdfile2 *T1);
 void local_filter_T1(dpdfile2 *T1);
 
 void L1_build(const struct L_Params& L_params) {
