+++ psi4/src/psi4/cc/cclambda/cc2_L1.cc
@@ -35,7 +35,7 @@
 #include "psi4/libqt/qt.h"
 #include "MOInfo.h"
 #include "Params.h"
-#include "cclambda.h"
+#include "Local.h"
 #define EXTERN
 #include "globals.h"
